requirements.txt
APScheduler==3.6.3
asgiref==3.11.1
cachetools==4.2.2
certifi==2026.2.25
Django==5.2.12
django-environ==0.13.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-telegram-bot==13.15
pytz==2026.1.post1
six==1.17.0
sqlparse==0.5.5
tornado==6.1
tzlocal==5.3.1
